Parser Declare Fce

+++ tests/Parser/Test_02_Define_Function.bas
@@ -1,4 +1,4 @@
-Declare Function factorial (n As Integer) As Integer
+Declare Function factorial (n As Double) As Integer
 Function factorial (n As Integer) As Integer
 	Return 0
 End Function
